Parallaxing now works properly
Modified the script as well as added two new images. One for stars in
the furthest back and one for the stars that are closer.

I also rearranged the canvas so that the scaling makes sense and is not
dependent on the canvas but on the actual metrics in the project.